move Miningmachine to separated class
